+++ operator-common/src/main/java/io/strimzi/operator/common/operator/resource/AbstractScalableResourceOperator.java
@@ -30,11 +30,7 @@ public abstract class AbstractScalableResourceOperator<C extends KubernetesClien
         extends AbstractReadyResourceOperator<C, T, L, R> {
 
     public static final String ANNO_STRIMZI_IO_GENERATION = Annotations.STRIMZI_DOMAIN + "generation";
-    @Deprecated
-    public static final String ANNO_OP_STRIMZI_IO_GENERATION = "operator." + Annotations.STRIMZI_DOMAIN + "generation";
     public static final String ANNO_STRIMZI_IO_DELETE_POD_AND_PVC = Annotations.STRIMZI_DOMAIN + "delete-pod-and-pvc";
-    @Deprecated
-    public static final String ANNO_OP_STRIMZI_IO_DELETE_POD_AND_PVC = "operator." + Annotations.STRIMZI_DOMAIN + "delete-pod-and-pvc";
 
     private final Logger log = LogManager.getLogger(getClass());